requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2==2.9.3